+++ multiboot2/Changelog.md
@@ -1,5 +1,12 @@
 # CHANGELOG for crate `multiboot2`
 
+## 0.14.2 (2023-03-17)
+- documentation fixes
+- `MbiLoadError` now implements `Display`
+- Added the `unstable` feature, which enables nightly-only functionality.
+  With this feature, `MbiLoadError` now implements `core::error::Error` and can
+  be used with `anyhow::Result` for example.
+
 ## 0.14.1 (2023-03-09)
 - fixed the calculation of the last area of the memory map tag ([#119](https://github.com/rust-osdev/multiboot2/pull/119))
   (Previously, iterating the EFI Memory map resulted in a superfluous entry as it ran over the next tag)